+++ CHANGELOG.md
@@ -206,3 +206,13 @@ Fix of `CalculatedDateTime#asPastFor` - now it will decrease time even if source
 * Add extension `DateTime#withoutTimeZoneOffset`
 * Fix of `GarbageCollector` time detection
 * `CalculatedDateTime#asFutureFor` will guarantee return future datetime even if `source` is equal to generated `now`
+
+## 0.7.0
+
+* Add `CallbackQueryReceivers` package with receivers interface and implementations
+* Rewrite rating receivers to use `SafeCallbackQueryReceiver`
+* Now rating receivers will automatically register message for post as likes message if have no registered
+* `TimerTriggerStrategy` now do not using synchronisations
+* Add `Iterable<CalculatedDateTime>#nearDateTime`
+* `TimerTriggerStrategy` config now receive time as parameter
+* Now `DatabaseConfig` is independent class which can connect to Database